pyproject.toml
[project]
name = "vllm-bootstrap"
version = "0.1.0"
description = "FastAPI control plane for launching and managing vLLM jobs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.6.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]
vllm = [
    "vllm>=0.6.0",
]

[project.scripts]
vllm-bootstrap = "vllm_bootstrap.__main__:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vllm_bootstrap"]

[tool.ruff]
exclude = ["vllm_bootstrap/generated"]

[tool.ruff.lint]
select = ["E402"]

[tool.pytest.ini_options]
testpaths = ["vllm_bootstrap/__tests__"]
